Fix bug with fast tab switching
Switching between 2 tabs back and forth toooo fast used to cause one app to get on the wrong navigation order.

Now if it's very fast they still still at times might fail switching, but at least the order stays correct.

I fixed it by proactively adding the tab to history right after the notification for switching the tab was received by Safari, instead of waiting for the tab to switch which used to trigger the update which adds it to history.