Fix bug with fast tab switching AGAIN
Initially i fixed it with this commit:
https://github.com/kopyl/safari-tab-switcher/commit/d53b3fc3a12e362f61041484fbc030dc2ef75771

But it introduced a bug where a user was unable to switch from a second (true order, not history) tab to any else.

Then after some optimizations it turned out that there is no such bug anymore.
So I removed this proactive tab saving in this commit:
https://github.com/kopyl/safari-tab-switcher/commit/20a63826886bf0fe1af5d6a4a1009f825112f425

Then i after few commits it turned out there is no such bug only when there are just a few tabs open. When there are quite many, the bug starts appearing again.

So in this commit i brought back the proactive new tab saving (the tab which a user just switched to).