Fix tabs panel sometimes not appearing
It was happening because sometime the app was unable to catch up to user.
Maybe because of race conditions, I'm not 100% sure.

So now whenever a Safari receives a message from app to open the tab, i immediately send a message from extension to the app asking the app to make sure it's tracking the app HotKey.