Fix typing in a tabs search field
Trying to delete characters not from the end of a search query would result in characters removed from the end of it.

The same with appending. When the cursor was not at the end of the search query and users would type anything, the characters were inserted at the of the search query instead of inserting exactly at the position of a cursor (where a user placed that cursor).